Make `ResidueIter` yield structs instead of references
This makes copies of positions and velocities, but also makes it so much easier to work with the borrow checker. By returning references, the object which `iter_residues` was called on would not be released until all these references would have been used up, which was cumbersome and unintuitive. By returning clones instead of references, the lock is removed.

Consider adding a separated method to get references?